requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
boto3==1.7.71
botocore==1.10.71
cffi==1.11.5
chalice==1.6.0
click==6.7
cryptography==2.3
docutils==0.14
enum-compat==0.0.2
flake8==3.5.0
idna==2.7
jmespath==0.9.3
mccabe==0.6.1
peewee==3.1.0
peewee-validates==1.0.7
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
typing==3.6.4
validate-email==1.3
